Added route popup and improved table style
